+++ lib/flowr_mvvm_support.dart
@@ -1,0 +1,2 @@
+export 'package:flowr/src/mvvm/mixin/change_notifier.dart';
+export 'package:provider/provider.dart' hide ReadContext;